refactor: add ProGuard rules for plaid connect
